Improved chart hover highlighting to update value axis and set a background for the legend items so the hover affects the entire item Changed Avalonia TextEditor to only fill the height required

+++ Libraries/Atlas.Tabs.Test/DataGrid/Tab - Test - Grid - Column - Types.cs
@@ -17,7 +17,7 @@ namespace Atlas.Tabs.Test.DataGrid
 					var testItem = new TestItem()
 					{
 						Integer = i,
-						Long = (long)i * (long)int.MaxValue,
+						Long = (long)i * int.MaxValue,
 						DateTime = new DateTime(DateTime.Now.Ticks + i),
 						TimeSpan = TimeSpan.FromHours(i),
 						Bool = (i % 2 == 1),
